business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/VfoOperation.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class VfoOperation implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+G ");
        messageString.append(command.getParameter("Mem/VFO Op"));
        messageString.append("\n");
        return messageString;
    }
}
